hacking: Add script to extract sysfs cpu information from a Linux target
The information is ouput as a single JSON file, and is meant to be used in
unit tests for CPU facts parsing on Linux.

An alternative would be to simply copy the sysfs directory structure, but it does not work well:

- rsync refuses to read files in sysfs, because it is confused about the
  real size being much smaller than the advertised size (4 KB)

- tar can read these files, but it pads them to 4 KB with \0, which wastes
  a lot of disk space

- scp correctly handles these files, but it's painfully slow over a
  high-RTT link

- in any case, the directory structure has more than 1000 files, which
  would impose unnecessary strains on git and the local filesystem

Dumping this data as JSON is simple and space-efficient (it takes around
10 KB - 100 KB depending on the host).  However, it also means that we
cannot extract binary data.  So far this isn't a problem.